feat(web-ui): appearance settings tab, nav layout, chat input styles
- Extract AppearanceConfig from Basics with i18n (en/zh-CN/zh-TW)

- NavPanel/sessions/workspaces and shell menu state tweaks

- Add fixedPopoverViewport helper; refresh ChatInput styles
