+++ src/web-ui/src/app/components/NavPanel/NavPanel.scss
@@ -895,6 +895,9 @@ $_section-header-height: 24px;
     display: inline-block;
     width: fit-content;
     max-width: min(440px, calc(100vw - 32px));
+    max-height: min(560px, calc(100vh - 16px));
+    overflow-x: hidden;
+    overflow-y: auto;
     vertical-align: top;
     box-sizing: border-box;
     padding: $size-gap-1 0;
