+++ src/web-ui/src/app/components/NavPanel/sections/workspaces/WorkspaceListSection.scss
@@ -879,6 +879,9 @@
     width: max-content;
     min-width: 180px;
     max-width: min(320px, calc(100vw - 24px));
+    max-height: min(560px, calc(100vh - 16px));
+    overflow-x: hidden;
+    overflow-y: auto;
     padding: $size-gap-1 0;
     border-radius: $size-radius-base;
     background: var(--color-bg-elevated, #1e1e22);
