+++ src/web-ui/src/app/scenes/settings/SettingsScene.tsx
@@ -16,6 +16,7 @@ import McpToolsConfig from '../../../infrastructure/config/components/McpToolsCo
 import AcpAgentsConfig from '../../../infrastructure/config/components/AcpAgentsConfig';
 import EditorConfig from '../../../infrastructure/config/components/EditorConfig';
 import BasicsConfig from '../../../infrastructure/config/components/BasicsConfig';
+import AppearanceConfig from '../../../infrastructure/config/components/AppearanceConfig';
 import ReviewConfig from '../../../infrastructure/config/components/ReviewConfig';
 
 const KeyboardShortcutsTab = lazy(() => import('./components/KeyboardShortcutsTab'));
@@ -39,6 +40,7 @@ const SettingsScene: React.FC = () => {
 
   switch (activeTab) {
     case 'basics':           Content = BasicsConfig;         break;
+    case 'appearance':       Content = AppearanceConfig;     break;
     case 'models':           Content = AIModelConfig;        break;
     case 'session-config':   Content = SessionConfig;        break;
     case 'review':           Content = ReviewConfig;         break;
