+++ src/web-ui/src/infrastructure/config/components/index.ts
@@ -5,6 +5,7 @@ export { default as MCPResourceBrowser } from './MCPResourceBrowser';
 export { default as McpToolsConfig } from './McpToolsConfig';
 export { default as EditorConfig } from './EditorConfig';
 export { default as BasicsConfig } from './BasicsConfig';
+export { default as AppearanceConfig } from './AppearanceConfig';
 
 
 export { default as DefaultModelConfig } from './DefaultModelConfig';
