+++ src/web-ui/src/locales/en-US/settings/appearance.json
@@ -1,0 +1,4 @@
+{
+  "title": "Appearance",
+  "subtitle": "Interface language, theme, and global font size"
+}
