+++ src/web-ui/src/locales/zh-CN/settings.json
@@ -8,7 +8,8 @@
     "searchClear": "清除搜索",
     "beta": "Beta",
     "tabDescriptions": {
-      "basics": "语言、主题、日志、终端 Shell 与开机启动。",
+      "basics": "日志、终端 Shell、通知与开机启动。",
+      "appearance": "语言、主题与界面字体大小。",
       "models": "AI 模型、API 密钥、供应商与代理。",
       "sessionConfig": "会话行为、工具与超时。",
       "review": "深度审核策略、审核员模型与额外审核 Sub-Agent。",
@@ -26,6 +27,7 @@
     },
     "tabs": {
       "basics": "基础",
+      "appearance": "外观",
       "models": "模型",
       "sessionConfig": "会话配置",
       "review": "审核",
