+++ src/web-ui/src/locales/zh-CN/settings/appearance.json
@@ -1,0 +1,4 @@
+{
+  "title": "外观",
+  "subtitle": "界面语言、主题与全局字体大小"
+}
