+++ src/web-ui/src/locales/zh-CN/settings/basics.json
@@ -1,6 +1,6 @@
 {
   "title": "基础",
-  "subtitle": "界面语言、主题、日志、终端与登录启动",
+  "subtitle": "日志、终端、通知与登录时启动",
   "appearance": {
     "title": "外观",
     "hint": "界面语言与视觉主题",
