+++ src/web-ui/src/locales/zh-TW/settings/appearance.json
@@ -1,0 +1,4 @@
+{
+  "title": "外觀",
+  "subtitle": "介面語言、主題與全域字體大小"
+}
