+++ src/web-ui/src/locales/zh-TW/settings/basics.json
@@ -1,6 +1,6 @@
 {
   "title": "基礎",
-  "subtitle": "界面語言、主題、日誌、終端與登錄啟動",
+  "subtitle": "日誌、終端、通知與登入時啟動",
   "appearance": {
     "title": "外觀",
     "hint": "界面語言與視覺主題",
